+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonColorDialog.cs
@@ -31,7 +31,7 @@ namespace Krypton.Toolkit
         {
             get => _commonDialogHandler.Icon;
             set => _commonDialogHandler.Icon = value;
-        } 
+        }
 
         /// <summary>
         /// Changes the default Icon to Developer set
@@ -83,17 +83,53 @@ namespace Krypton.Toolkit
         protected override IntPtr HookProc(IntPtr hWnd, int msg, IntPtr wparam, IntPtr lparam)
         {
             var (handled, retValue) = _commonDialogHandler.HookProc(hWnd, msg, wparam, lparam);
+            if (msg == PI.WM_.INITDIALOG
+                && !FullOpen
+                )
+            {
+                // Find the Static colour set 000002D0
+                var clrColourBox = _commonDialogHandler.Controls.FirstOrDefault(ctl => ctl.DlgCtrlId == 0x000002D0);
+                var rcClient = clrColourBox.WinInfo.rcClient;
+                var lpPoint = new PI.POINT(rcClient.left, rcClient.top);
+                PI.ScreenToClient(hWnd, ref lpPoint);
+                clrColourBox.ClientLocation = new Point(lpPoint.X, lpPoint.Y);
+                clrColourBox.Size = new Size(rcClient.right - rcClient.left, rcClient.bottom - rcClient.top);
+
+                // Find the bottom of the OK button (Might not have OK text !!) 00000001
+                var btnOk = _commonDialogHandler.Controls.FirstOrDefault(ctl => ctl.DlgCtrlId == 0x00000001);
+
+                // Now adjust the size so that it the correct display on "All" supported OS's
+                // https://github.com/Krypton-Suite/Standard-Toolkit/issues/415
+                Size toolBoxSize = _commonDialogHandler._toolBox.ClientSize;
+                toolBoxSize.Width = clrColourBox.Size.Width + 2 * clrColourBox.ClientLocation.X;
+                toolBoxSize.Height = btnOk.ClientLocation.Y + btnOk.Size.Height *3 /2;
+
+                _commonDialogHandler._toolBox.ClientSize = toolBoxSize;
+            }
+
             if (!handled
                 && (msg == PI.WM_.WINDOWPOSCHANGING)
                 && _commonDialogHandler.EmbeddingDone
             )
             {
                 PI.WINDOWPOS pos = (PI.WINDOWPOS)PI.PtrToStructure(lparam, typeof(PI.WINDOWPOS));
-                if ( !pos.flags.HasFlag(PI.SWP_.NOSIZE)
+                if (!pos.flags.HasFlag(PI.SWP_.NOSIZE)
                     && (pos.hwnd == hWnd)
                     )
                 {
-                    handled = _commonDialogHandler.SetNewPosAndClientSize(new Point(pos.x, pos.y), new Size(pos.cx, pos.cy));
+                    var newSize = new Size(pos.cx, pos.cy);
+                    if (!FullOpen
+                        && newSize.Width > _commonDialogHandler._toolBox.Size.Width
+                    )
+                    {
+                        // Need to fiddle the width and height to workaround the Magic hidden "&d" button
+                        // https://github.com/Krypton-Suite/Standard-Toolkit/issues/416
+                        if (!ShowIcon)
+                            newSize.Width -= 16;
+                        newSize.Height -= 44;
+                    }
+
+                    handled = _commonDialogHandler.SetNewPosAndClientSize(new Point(pos.x, pos.y), newSize);
                     if (!handled)
                     {
                         pos.flags |= PI.SWP_.NOSIZE;
@@ -103,7 +139,7 @@ namespace Krypton.Toolkit
                     retValue = IntPtr.Zero;
                 }
             }
-            return handled ? retValue :base.HookProc(hWnd, msg, wparam, lparam);
+            return handled ? retValue : base.HookProc(hWnd, msg, wparam, lparam);
         }
 
     }